Wed, Sep 6, 2023, 1:42 PM +01:00

+++ section-exercise1.c
@@ -9,6 +9,6 @@ int main(void)
   int age = get_int("how hold are you? ");
   printf("You are %i years old\n", age);
 
-  int phoneNumber = get_long("Give me your phone number: ");
+  long phoneNumber = get_long("Give me your phone number: ");
   printf("Your phone number is %li\n", phoneNumber);
 }
